Refactor native Rust module into focused source files and update docs
